Add audit logs endpoint
